added Task_2

+++ AutomationCourse_HomeWork/Task_1.cs
@@ -8,11 +8,11 @@ namespace AutomationCourse_HomeWork
         {
             Console.WriteLine("Please add three arguments for the futher operations:");
             Console.Write("x = ");
-            int x = Convert.ToInt32(Console.ReadLine());
+            double x = Convert.ToInt32(Console.ReadLine());
             Console.Write("y = ");
-            int y = Convert.ToInt32(Console.ReadLine());
+            double y = Convert.ToInt32(Console.ReadLine());
             Console.Write("z = ");
-            int z = Convert.ToInt32(Console.ReadLine());
+            double z = Convert.ToInt32(Console.ReadLine());
 
             double result1 = (x + y) * (1 + Math.Sqrt(z));
             Console.WriteLine($"The result of the first statement: {result1}");
